chore: relocate misplaced files and fix stale docs
