--- .cargo/config.toml
@@ -1,7 +1,0 @@
-[target.armv7-unknown-linux-gnueabihf]
-linker = "arm-linux-gnueabihf-gcc"
-
-[env]
-CC_armv7_unknown_linux_gnueabihf = "arm-linux-gnueabihf-gcc"
-CXX_armv7_unknown_linux_gnueabihf = "arm-linux-gnueabihf-g++"
-AR_armv7_unknown_linux_gnueabihf = "arm-linux-gnueabihf-ar"
