+++ docs/img/checked_in.png
