+++ docs/img/checked_out.png
