+++ docs/img/placeholder_nfc_scan.png
