+++ docs/img/toilet_icon.png
