+++ src/types/network.ts
@@ -1,0 +1,6 @@
+export interface NetworkStatusData {
+  isOnline: boolean;
+  responseTime: number;
+  lastChecked: number;
+  quality: 'online' | 'poor' | 'offline';
+}
